task4_automation_orchestrator_python/requirements.txt
python-dotenv
google-generativeai
schedule
requests

